# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dientes Fijos Medellín CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: treatment
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientBody"
      responses:
        "200":
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /treatments:
    get:
      operationId: listTreatments
      tags: [treatments]
      summary: List all treatments
      responses:
        "200":
          description: List of treatments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treatment"
    post:
      operationId: createTreatment
      tags: [treatments]
      summary: Create treatment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreatmentBody"
      responses:
        "201":
          description: Treatment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
  /treatments/{id}:
    put:
      operationId: updateTreatment
      tags: [treatments]
      summary: Update treatment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTreatmentBody"
      responses:
        "200":
          description: Treatment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
    delete:
      operationId: deleteTreatment
      tags: [treatments]
      summary: Delete treatment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        age:
          type: integer
        treatment:
          type: string
        status:
          type: string
          enum: [new, interested, scheduled, attended, in_treatment, completed]
        lastVisit:
          type: string
          format: date
          nullable: true
        nextAppointment:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, phone, status, createdAt]
    CreatePatientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        age:
          type: integer
        treatment:
          type: string
        status:
          type: string
          enum: [new, interested, scheduled, attended, in_treatment, completed]
        notes:
          type: string
      required: [name, phone]
    UpdatePatientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        age:
          type: integer
        treatment:
          type: string
        status:
          type: string
          enum: [new, interested, scheduled, attended, in_treatment, completed]
        notes:
          type: string
...
    Treatment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        duration:
          type: integer
        active:
          type: boolean
      required: [id, name, price, duration, active]
    CreateTreatmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        duration:
          type: integer
        active:
          type: boolean
      required: [name, price, duration]
    UpdateTreatmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        duration:
          type: integer
        active:
          type: boolean
...